extract score class
