+++ src/main/java/tetris/Controller.java
@@ -1,6 +1,5 @@
 package tetris;
 
-import java.util.Observable;
 import java.util.Observer;
 
 import tetris.tetromino.AbstractTetromino;
@@ -18,40 +17,6 @@ import logging.Logger;
 
 public class Controller {
 
-    private class Score extends Observable {
-        private int score = 0;
-        private int level = 0;
-        private int[] pointsPerRow = { 0, 40, 50, 100, 300 };
-
-        /**
-         * resets the score to 0.
-         */
-        public void reset() {
-            score = 0;
-            level = 0;
-        }
-
-        /**
-         * increases the score by add.
-         * 
-         * @param rowsCleared
-         *            add add points to the score
-         */
-        public void add(int rowsCleared) {
-            score += rowsCleared * pointsPerRow[rowsCleared];
-            level += rowsCleared;
-            setChanged();
-            notifyObservers(score);
-        }
-
-        /**
-         * @return the current level that has been reached.
-         */
-        public int getLevel() {
-            return level / 10;
-        }
-    }
-
     private Score score;
     private View ui;
     private Grid grid;
@@ -60,7 +25,6 @@ public class Controller {
     private boolean gameOver = false;
     private EventHandler<ActionEvent> onTick = event -> lowerTetromino();
     private Tick timer = new Tick(onTick);
-
     private Settings settings;
 
     /**
